Fix iCRC type to make it 4-byte on both 32-bit and 64-bit platforms

+++ include/AddonFormat.h
@@ -22,7 +22,7 @@ namespace Addon
 	{
 		Bootil::BString	strName;
 		long long		iSize;
-		unsigned long	iCRC;
+		unsigned int	iCRC;
 		unsigned int	iFileNumber;
 		long long		iOffset;
